Contemplate master/slave resources in check_resource_state
Note that this is just a small incremental fix so we can also
grep for "Master" in case we want the state to be 'started'.
The reason for this is that multi_master resource won't show
up with "Started" in pcs status:

Master/Slave Set: galera-master [galera]
    galera     (ocf::heartbeat:galera):        Master overcloud-controller-1
    galera     (ocf::heartbeat:galera):        Master overcloud-controller-2
    galera     (ocf::heartbeat:galera):        Master overcloud-controller-0
    Masters: [ overcloud-controller-0 overcloud-controller-1 overcloud-controller-2 ]

So for now we just also grep for 'master'. This function is in
dire need of a substantial rewrite though that will take the
following into account:
- Some master/slave resources need to be Master on all nodes (e.g.
  galera)
- Some master/slave resources need to be Master on one node and slave on
  all others (e.g. redis)
- We need to take into account that with composable HA we will see
  'Stopped' on all nodes where the resource is not supposed to be run

Change-Id: I1971f6ebd5cd4da6903ecebe1acd009d2437a7b7
Related-Bug: #1668372